Split single router into multiple
