+++ api/routes/open/chefs-router.js
@@ -1,0 +1,10 @@
+import express from 'express';
+import { submitRequestHandler } from '../../controllers/chefs-api-controller.js';
+
+const router = express.Router();
+
+// CHEFS endpoints
+router.route('/requests')
+  .post(submitRequestHandler);
+
+export default router;
