+++ api/routes/open/index.js
@@ -1,0 +1,9 @@
+import healthRouter from './health-router.js';
+import chefsRouter from './chefs-router.js';
+
+const openRouter = {
+  healthRouter,
+  chefsRouter
+};
+
+export default openRouter;
