+++ api/routes/protected/index.js
@@ -1,0 +1,7 @@
+import keycloakTest from './keycloak-test-router.js';
+
+const protectedRouter = {
+  keycloakTest
+};
+
+export default protectedRouter;
